Move train_valid_split to util
